Hans' xmame-0.37b6.1 relase unix tree. Please check if any recent changes got overwritten.

+++ src/unix/osd_cpu.h
@@ -51,7 +51,7 @@ typedef unsigned int       UINT32;
 typedef unsigned long long UINT64;
 
 /* Combine to 32-bit integers into a 64-bit integer */
-#define COMBINE_64_32_32(A,B)     ((((UINT64)(A))<<32) | (B))
+#define COMBINE_64_32_32(A,B)     ((((UINT64)(A))<<32) | (UINT32)(B))
 #define COMBINE_U64_U32_U32(A,B)  COMBINE_64_32_32(A,B)
 
 /* Return upper 32 bits of a 64-bit integer */
